Gut ProjectContext
Since it doesn't do anything other than resolving references.
