+++ src/project-context.h
@@ -9,29 +9,33 @@
 class PB_SHARED_EXPORT ProjectContext : public QObject
 {
     Q_OBJECT
-    Q_PROPERTY(QQmlListProperty<Attribute> all_attributes READ all_attributes)
-    Q_PROPERTY(QQmlListProperty<BonusSource> all_bonus_sources READ all_bonus_sources)
 
 public:
-    using AttributeList = QQmlListProperty<Attribute>;
-    using BonusSourceList = QQmlListProperty<BonusSource>;
+    using ResourceList = QQmlListProperty<Resource>;
 
     explicit ProjectContext(QObject *parent = 0);
     ~ProjectContext();
 
-    AttributeList all_attributes();
-    BonusSourceList all_bonus_sources();
-
     void setCharacterRoot(QObject *root);
-
-    template <typename T>
-    static void PB_SHARED_EXPORT populateInstancesOf(QObject *obj, QList<T*> &res);
-
     Resource::List allResources();
 
+    template <typename T>
+    static void populateInstancesOf(QObject *obj, QList<T*> &res);
+
 private:
-    BonusSource::List m_bonusSrcList;
     Resource::List m_resources;
 };
+
+template <typename T>
+void ProjectContext::populateInstancesOf(QObject *obj, QList<T*> &res) {
+    auto cast = qobject_cast<T*>(obj);
+    if (cast) {
+        res.append(cast);
+    }
+
+    for (QObject *child : obj->children()) {
+        populateInstancesOf(child, res);
+    }
+}
 
 #endif // PROJECTCONTEXT_H
